Разделяет utils на модули
