+++ src/utils/film.js
@@ -1,0 +1,7 @@
+import dayjs from 'dayjs';
+
+const humanizeDate = (date, template = 'D MMMM YYYY') => {
+  return dayjs(date).format(template);
+};
+
+export { humanizeDate };
